Develop 8888

+++ effect.cpp
@@ -700,7 +700,7 @@ int32 effect::get_value(uint32 extraargs) {
 		return res;
 	} else {
 		pduel->lua->params.clear();
-		return (int32)value;
+		return value;
 	}
 }
 int32 effect::get_value(card* pcard, uint32 extraargs) {
@@ -711,7 +711,7 @@ int32 effect::get_value(card* pcard, uint32 extraargs) {
 		return res;
 	} else {
 		pduel->lua->params.clear();
-		return (int32)value;
+		return value;
 	}
 }
 int32 effect::get_value(effect* peffect, uint32 extraargs) {
@@ -722,7 +722,7 @@ int32 effect::get_value(effect* peffect, uint32 extraargs) {
 		return res;
 	} else {
 		pduel->lua->params.clear();
-		return (int32)value;
+		return value;
 	}
 }
 void effect::get_value(uint32 extraargs, std::vector<lua_Integer>& result) {
@@ -751,7 +751,7 @@ void effect::get_value(effect* peffect, uint32 extraargs, std::vector<lua_Intege
 		pduel->lua->get_function_value(value, 2 + extraargs, result);
 	} else {
 		pduel->lua->params.clear();
-		result.push_back((int32)value);
+		result.push_back(value);
 	}
 }
 int32 effect::get_integer_value() {
@@ -764,7 +764,7 @@ int32 effect::check_value_condition(uint32 extraargs) {
 		return res;
 	} else {
 		pduel->lua->params.clear();
-		return (int32)value;
+		return value;
 	}
 }
 void* effect::get_label_object() {
